@@ -32,7 +32,7 @@ class BackupFactoryTest extends \PHPUnit\Framework\TestCase
      */
     protected $_data;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->_data = [
             'id' => '1385661590_snapshot',
@@ -46,8 +46,8 @@ class BackupFactoryTest extends \PHPUnit\Framework\TestCase
             $this->at(0)
         )->method(
             'getIterator'
-        )->will(
-            $this->returnValue(new \ArrayIterator([new \Magento\Framework\DataObject($this->_data)]))
+        )->willReturn(
+            new \ArrayIterator([new \Magento\Framework\DataObject($this->_data)])
         );
 
         $this->_backupModel = $this->createMock(\Magento\Backup\Model\Backup::class);
@@ -59,8 +59,8 @@ class BackupFactoryTest extends \PHPUnit\Framework\TestCase
             'create'
         )->with(
             \Magento\Backup\Model\Fs\Collection::class
-        )->will(
-            $this->returnValue($this->_fsCollection)
+        )->willReturn(
+            $this->_fsCollection
         );
         $this->_objectManager->expects(
             $this->at(1)
@@ -68,8 +68,8 @@ class BackupFactoryTest extends \PHPUnit\Framework\TestCase
             'create'
         )->with(
             \Magento\Backup\Model\Backup::class
-        )->will(
-            $this->returnValue($this->_backupModel)
+        )->willReturn(
+            $this->_backupModel
         );
 
         $this->_instance = new \Magento\Backup\Model\BackupFactory($this->_objectManager);
@@ -80,26 +80,26 @@ class BackupFactoryTest extends \PHPUnit\Framework\TestCase
         $this->_backupModel->expects($this->once())
             ->method('setType')
             ->with($this->_data['type'])
-            ->will($this->returnSelf());
+            ->willReturnSelf();
 
         $this->_backupModel->expects($this->once())
             ->method('setTime')
             ->with($this->_data['time'])
-            ->will($this->returnSelf());
+            ->willReturnSelf();
 
         $this->_backupModel->expects($this->once())
             ->method('setName')
             ->with($this->_data['name'])
-            ->will($this->returnSelf());
+            ->willReturnSelf();
 
         $this->_backupModel->expects($this->once())
             ->method('setPath')
             ->with($this->_data['path'])
-            ->will($this->returnSelf());
+            ->willReturnSelf();
 
         $this->_backupModel->expects($this->once())
             ->method('setData')
-            ->will($this->returnSelf());
+            ->willReturnSelf();
 
         $this->_instance->create('1385661590', 'snapshot');
     }
